File ReStructure and CleanUp
- Reorganized the file structure.
- Cleaned up some files.
