+++ FirstProject/taskopedia/src/components/Student.jsx
@@ -3,7 +3,7 @@ import { faker } from "@faker-js/faker";
 
 function Student(props) {
   return (
-    <div className="container p-4">
+    <div className="col-4 p-1">
       <div className="row border">
         <div className="col-2">
           <img src={faker.image.avatar()} className="w-100 py-2" alt={props.name} />
