+++ FirstProject/taskopedia/src/components/layout/Footer.jsx
@@ -1,0 +1,9 @@
+import React from "react";
+
+function Footer() {
+  return (
+    <p style={{ backgroundColor: "black", color: "gray", marginTop: "0.625em", textAlign: "center" }}>Happy Coding!</p>
+  );
+}
+
+export default Footer;
