+++ FirstProject/taskopedia/src/scss/style.scss
@@ -1,8 +1,17 @@
+body {
+  background-color: black;
+}
+
 .heading1 {
-  color: brown;
+  background-color: black;
   font-size: large;
 }
 
 .h1 {
   color: chocolate;
+}
+
+.header-logo {
+  height: 2.1875em;
+  vertical-align: top;
 }
